Add memory, stack, timers and font
Allocate 4KB for the memory. Use a uint16 stack to
represent the stack and uint8 for the timers.
The font is an array added to the memory at
0x50. Decrement the timers when they are not 0
at 60hz. Use SDL3 for screen and sound.

+++ main.cpp
@@ -35,7 +35,7 @@ uint8_t font[] = {
 struct Machine {
     uint8_t* memory = new uint8_t[0x1000]{}; // 4KB initialized to 0
 
-    stack<uint8_t> pila; // stack
+    stack<uint16_t> pila; // stack
 
     uint8_t sound_timer = 0;
     uint8_t delay_timer = 0;
